admin-service/requirements.txt
asgiref==3.3.4
Django==3.2.4
django-cors-headers==3.7.0
django-mysql==3.12.0
djangorestframework==3.12.4
pika==1.2.0
pytz==2021.1
sqlparse==0.4.1
mysqlclient==2.0.1